app/requirement.txt
flask
mysql-connector
dependency-injector